# perf: do less allocs in hasher